로그인 주체만 '완료' 뜨는거 최적화
