+++ src/components/GroupSpace/CheckList/Person/PList.jsx
@@ -14,6 +14,19 @@ const PList = ({ selectedName, selectedParentPlace }) => {
     const [isAddMode, setIsAddMode] = useState(false);
     const [text, setText] = useState("편집");
     const [trigger, setTrigger] = useState(0);
+    const [owner, setIsOwner] = useState("임시");
+
+    useEffect(() => {
+        const fetchOwner = async () => {
+            try {
+                const res = await axiosInstance.get("/mypage/info/");
+                setIsOwner(res.data.data.name);
+            } catch (error) {
+                console.error("로그인 주체 불러옴:", error);
+            }
+        };
+        fetchOwner();
+    }, []);
 
     useEffect(() => {
         const fetchData = async () => {
@@ -162,6 +175,7 @@ const PList = ({ selectedName, selectedParentPlace }) => {
                         item={item}
                         setTrigger={setTrigger}
                         selectedName={selectedName}
+                        owner={owner}
                     />
                 ))}
                 {isEditMode && (
